feat(pty): register commands, regenerate bindings, snapshot wire format
Four pty_* commands enter collect_commands! and the allowlist; bindings.ts
regenerated; insta snapshots pin the request arg shapes and the PtyError
envelope. build.rs drift check now sees both files agree.

Also narrows PtyId from u64 to u32 to satisfy Specta's BigInt-forbidden rule
(u64 cannot be exported to TypeScript without precision loss).

+++ src-tauri/src/pty/types.rs
@@ -7,7 +7,7 @@ use specta::Type;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Type)]
 #[serde(transparent)]
-pub struct PtyId(pub u64);
+pub struct PtyId(pub u32);
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, Type)]
 pub struct PtySize {
